feat: implementation of TS common viewContent, URL fragment & remote URL content (#2841)
* feat: initial implementation of TS viewContent & URL fragment.
- Add view content processing, to inject common view content
  when a view is loaded.
- Add URL hash/fragment location when loading the view.
* fix: getStyle instead of getTokenView
* fix: Add minimum TS schema version to support old versions
This allows adding new features to future schema versions
whilst maintaining backward compatibility with old schemas.

It adds a new static variable TOKENSCRIPT_MINIMUM_SCHEMA and
function isSchemaLessThanMinimum. This it then used to determine
if old definitions should be deleted. Old versions are now only
deleted if they don't meet the minimum, rather than not matching
the TOKENSCRIPT_CURRENT_SCHEMA.

* fix: namespace null check
* feat: TS - Add URL attribute functionality to load remote views

This change allows TokenScripts to use the url attribute of a view to
specify that the view should be loaded from a remote location such as IPFS.
For bigger applications, it becomes hard for the developer to embed all required
assets inside a small TokenScript; such as fonts, large graphics and other media.

This provides a different route that make it easier to make richer applications
while ensuring TokenScript TSML size remains small. It also demonstrates the
potential for existing DApps to be easily ported to TokenScript, provided
that Web3TokenView.java is updated to provide the full ethereum provider to the client.
Access to the full provider can be specified as a TokenScript view attribute,
adding an extra layer of security.